refactor: Question 도메인에 대해서 CQS(Command Query Separation) 적용
