--- src/main/java/com/codelion/animalcare/domain/doctorqna/service/QuestionService.java
@@ -1,131 +1,0 @@
-package com.codelion.animalcare.domain.doctorqna.service;
-
-import com.codelion.animalcare.domain.doctorqna.dto.request.QuestionSaveRequestDto;
-import com.codelion.animalcare.domain.doctorqna.dto.request.QuestionUpdateRequestDto;
-import com.codelion.animalcare.domain.doctorqna.dto.response.QuestionResponseDto;
-import com.codelion.animalcare.domain.doctorqna.repository.Question;
-import com.codelion.animalcare.domain.doctorqna.repository.QuestionLike;
-import com.codelion.animalcare.domain.doctorqna.repository.QuestionLikeRepository;
-import com.codelion.animalcare.domain.doctorqna.repository.QuestionRepository;
-import com.codelion.animalcare.domain.user.entity.Member;
-import com.codelion.animalcare.domain.user.entity.UserInfo;
-import com.codelion.animalcare.domain.user.service.MemberService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.security.Principal;
-import java.util.ArrayList;
-import java.util.List;
-
-@RequiredArgsConstructor
-@Service
-@Transactional(readOnly = true)
-public class QuestionService {
-
-    private final QuestionRepository questionRepository;
-    private final MemberService memberService;
-    private final QuestionLikeRepository questionLikeRepository;
-
-    @Transactional
-    public Long save(QuestionSaveRequestDto questionSaveRequestDto, Principal principal) {
-        Member member = memberService.findMemberByEmail(principal.getName());
-        return questionRepository.save(questionSaveRequestDto.toEntity(member)).getId();
-    }
-
-    @Transactional
-    public Long update(Long id, QuestionUpdateRequestDto questionUpdateRequestDto){
-        Question question = findQuestionByQuestionId(id);
-        question.update(questionUpdateRequestDto.getTitle(), questionUpdateRequestDto.getContent());
-
-        return id;
-    }
-
-
-    @Transactional(readOnly = true)
-    public QuestionResponseDto findById(Long id){
-        Question entity = findQuestionByQuestionId(id);
-        return new QuestionResponseDto(entity);
-    }
-
-    @Transactional
-    public int updateView(Long id) {
-        return this.questionRepository.updateView(id);
-    }
-
-    @Transactional(readOnly = true)
-    public Page<Question> findAll(int page, String type, String kw) {
-        List<Sort.Order> sortsList = new ArrayList<>();
-        sortsList.add(Sort.Order.desc("createdAt"));
-
-        Pageable pageable = PageRequest.of(page, 10, Sort.by(sortsList));
-
-        switch (type != null ? type : " ") {
-            case "title" -> {
-                return questionRepository.findByTitleContaining(kw, pageable);
-            }
-            case "content" -> {
-                return questionRepository.findByContentContaining(kw, pageable);
-            }
-            case "member" -> {
-                return questionRepository.findByMemberContaining(kw, pageable);
-            }
-            case "titleAndContent" -> {
-                return questionRepository.findByTitleOrContent(kw, pageable);
-            }
-            default -> {
-                return questionRepository.findAll(pageable);
-            }
-        }
-
-    }
-
-    @Transactional
-    public void delete(Long id) {
-        Question question = findQuestionByQuestionId(id);
-        questionRepository.delete(question);
-    }
-
-    public boolean questionAuthorized(Long id, Principal principal){
-        Question question = findQuestionByQuestionId(id);
-
-        if(question.getMember().getEmail().equals(principal.getName())) {
-            return false;
-        }
-
-        return true;
-    }
-
-    public boolean findLike(Long id, UserInfo user) {
-        return questionLikeRepository.existsByQuestion_IdAndMember_Id(id, user.getId());
-    }
-
-    @Transactional(readOnly = false)
-    public boolean saveLike(Long id, UserInfo user){
-
-        if(!findLike(id, user)) {
-
-            Question question = questionRepository.findById(id).orElseThrow(IllegalArgumentException::new);
-            QuestionLike questionLike = new QuestionLike(user, question);
-            questionLikeRepository.save(questionLike);
-            questionRepository.plusLike(id);
-
-            return true;
-        }
-
-        questionLikeRepository.deleteByQuestion_IdAndMember_Id(id, user.getId());
-        questionRepository.minusLike(id);
-
-        return false;
-    }
-
-    public Question findQuestionByQuestionId(Long id) {
-        return questionRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("질문이 존재하지 않습니다."));
-    }
-
-}
-
